+++ src/test/java/playground/tests/exceptions/JupiterExecutableInvokerError.java
@@ -1,0 +1,18 @@
+package playground.tests.exceptions;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import ru.vyarus.spock.jupiter.support.ExecutableInvokerUsage;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 02.09.2022
+ */
+@Disabled
+@ExtendWith(ExecutableInvokerUsage.class)
+public class JupiterExecutableInvokerError {
+
+    @Test
+    public void test(){}
+}